Maintain a list of tracks similar to artists
Found a bug where when you select previous artists, it does not
update their tracks. This is because we aren't storing track history!
Rather than implement another undo-type functionality, I'm going to
take the same route as artists, and maintain a list of all seen
tracks.

A new prop is needed, `selectedArtistId`, so that we can derive which
of the tracks in our list should be displayed.

Some other fundamental changes are needed, specifically around handling
transitions to previous artists, and hard refreshes on URLs.